worker/pyproject.toml
[project]
name = "axon-worker"
version = "1.0.0"
description = "AXON audit worker — Playwright scraper + scoring + enrichment + PDF rendering."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "playwright>=1.46.0",
    "playwright-stealth>=1.0.6",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "asyncpg>=0.29.0",
    "structlog>=24.4.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-tasks>=2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-playwright>=0.5.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
